+++ test/functional/android/android/device_test.rb
@@ -271,12 +271,6 @@ class AppiumLibCoreTest
         assert @driver.get_display_density.positive?
       end
 
-      def test_keyevent
-        skip('Because only for Selendroid')
-        # http://developer.android.com/reference/android/view/KeyEvent.html
-        assert @driver.keyevent(176)
-      end
-
       def test_press_keycode
         # http://developer.android.com/reference/android/view/KeyEvent.html
         result = @driver.press_keycode(176) # it does not raise error