+++ lib/appium_lib_core/common/command.rb
@@ -213,8 +213,6 @@ module Appium
         hide_keyboard: [:post, 'session/:session_id/appium/device/hide_keyboard'],
         press_keycode: [:post, 'session/:session_id/appium/device/press_keycode'],
         long_press_keycode: [:post, 'session/:session_id/appium/device/long_press_keycode'],
-        # keyevent is only for Selendroid
-        keyevent: [:post, 'session/:session_id/appium/device/keyevent'],
         push_file: [:post, 'session/:session_id/appium/device/push_file'],
         pull_file: [:post, 'session/:session_id/appium/device/pull_file'],
         pull_folder: [:post, 'session/:session_id/appium/device/pull_folder'],