+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Read `release_notes.md` for commit level details.
         - test/functional/ios/webdriver/w3c_actions_test.rb
         - test/functional/common_w3c_actions.rb
         - https://www.selenium.dev/documentation/support_packages/mouse_and_keyboard_actions_in_detail/
+- Removed Selendroid related methods
 
 ## [4.7.1] - 2021-09-26