+++ CHANGELOG.md
@@ -29,6 +29,13 @@ Read `release_notes.md` for commit level details.
   - Can give `driver` instance as its block variable
 
 ### Deprecations
+- `Appium::Core::TouchAction` and `Appium::Core::MultiTouch` are deprecated
+    - Please use W3C actions instead http://appium.io/docs/en/commands/interactions/actions/
+    - More working examples:
+        - test/functional/android/webdriver/w3c_actions_test.rb
+        - test/functional/ios/webdriver/w3c_actions_test.rb
+        - test/functional/common_w3c_actions.rb
+        - https://www.selenium.dev/documentation/support_packages/mouse_and_keyboard_actions_in_detail/
 - Removed Selendroid related methods
 
 ## [4.7.1] - 2021-09-26